+++ packages/core/src/apex/ApexClassFetcher.ts
@@ -1,5 +1,5 @@
 import { Connection } from "@salesforce/core";
-import QueryHelper from "./QueryHelper";
+import QueryHelper from "../queryHelper/QueryHelper";
 
 
 export default class ApexClassFetcher {
