+++ packages/core/src/sfpcommands/package/CreateSourcePackageImpl.ts
@@ -5,7 +5,7 @@ import ProjectConfig from "../../project/ProjectConfig";
 import { EOL } from "os";
 import * as fs from "fs-extra";
 import path = require("path");
-import  { ApexSortedByType, FileDescriptor } from "../../parser/ApexTypeFetcher";
+import  { ApexSortedByType, FileDescriptor } from "../../apex/parser/ApexTypeFetcher";
 import SFPStatsSender from "../../stats/SFPStatsSender";
 import PackageEmptyChecker from "../../package/PackageEmptyChecker";
 import SFPPackage  from "../../package/SFPPackage";
